tests/requirements.txt
pytest==7.4.2
requests==2.31.0
pillow==10.0.1
numpy==1.24.3